get rid of unused matrix code

+++ src/Tests/DataType/SymmetricMatrixTest.cs
@@ -1,4 +1,4 @@
-// Copyright (C) 2011 Zeno Gantner
+// Copyright (C) 2011, 2012 Zeno Gantner
 //
 // This file is part of MyMediaLite.
 //
@@ -27,36 +27,36 @@ namespace Tests.DataType
 	{
 		[Test()] public void TestSymmetricity()
 		{
-			var matrix = new SymmetricMatrix<double>(5);
-			matrix[1, 3] = 1.0;
+			var matrix = new SymmetricMatrix<float>(5);
+			matrix[1, 3] = 1.0f;
 
 			Assert.AreEqual(1.0, matrix[3, 1]);
 		}
 
 		[Test()] public void TestIsSymmetric()
 		{
-			var matrix = new SymmetricMatrix<double>(5);
+			var matrix = new SymmetricMatrix<float>(5);
 			Assert.IsTrue(matrix.IsSymmetric);
 
-			matrix[1, 3] = 1.0;
+			matrix[1, 3] = 1.0f;
 			Assert.IsTrue(matrix.IsSymmetric);
 		}
 
 		[Test()] public void TestNumberOfRows()
 		{
-			var matrix = new SymmetricMatrix<double>(3);
+			var matrix = new SymmetricMatrix<float>(3);
 			Assert.AreEqual(3, matrix.NumberOfRows);
 		}
 
 		[Test()] public void TestNumberOfColumns()
 		{
-			var matrix = new SymmetricMatrix<double>(3);
+			var matrix = new SymmetricMatrix<float>(3);
 			Assert.AreEqual(3, matrix.NumberOfColumns);
 		}
 
 		[Test()] public void TestCreateMatrix()
 		{
-			var matrix1 = new SymmetricMatrix<double>(5);
+			var matrix1 = new SymmetricMatrix<float>(5);
 			var matrix2 = matrix1.CreateMatrix(4, 4);
 			Assert.IsInstanceOfType(matrix1.GetType(), matrix2);
 		}
